+++ src/RunTracy/Helpers/SlimRequestPanel.php
@@ -25,6 +25,7 @@ class SlimRequestPanel implements IBarPanel
 {
     private $app;
     private $ver;
+    private $icon;
 
     public function __construct($app = null, array $ver = [])
     {
@@ -34,11 +35,12 @@ class SlimRequestPanel implements IBarPanel
 
     public function getTab()
     {
-        return '<span title="Slim Http Request"><svg enable-background="new 0 0 64 64" height="64px" version="1.1" viewBox="0 0 64 64" width="64px" xml:space="preserve" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><g id="Layer_1"><g><circle cx="32" cy="32" fill="#77B3D4" r="32"/></g><g opacity="0.2"><path d="M49.529,33.855L35.259,15.71c-1.793-2.28-4.726-2.28-6.519,0L14.471,33.855    C12.679,36.135,13.439,38,16.161,38H24v12c0,2.209,1.791,4,4,4h8c2.209,0,4-1.791,4-4V38h7.839    C50.561,38,51.321,36.135,49.529,33.855z" fill="#231F20"/></g><g><path d="M40,48c0,2.209-1.791,4-4,4h-8c-2.209,0-4-1.791-4-4V24c0-2.209,1.791-4,4-4h8c2.209,0,4,1.791,4,4V48z" fill="#FFFFFF"/></g><g><path d="M16.161,36c-2.722,0-3.483-1.865-1.69-4.145L28.741,13.71c1.793-2.28,4.726-2.28,6.519,0l14.269,18.146    c1.793,2.28,1.032,4.145-1.69,4.145H16.161z" fill="#FFFFFF"/></g></g><g id="Layer_2"/></svg></span>';
+        $this->icon = '<svg enable-background="new 0 0 64 64" height="16px" version="1.1" viewBox="0 0 64 64" width="16px" xml:space="preserve" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><g id="Layer_1"><g><circle cx="32" cy="32" fill="#77B3D4" r="32"/></g><g opacity="0.2"><path d="M49.529,33.855L35.259,15.71c-1.793-2.28-4.726-2.28-6.519,0L14.471,33.855    C12.679,36.135,13.439,38,16.161,38H24v12c0,2.209,1.791,4,4,4h8c2.209,0,4-1.791,4-4V38h7.839    C50.561,38,51.321,36.135,49.529,33.855z" fill="#231F20"/></g><g><path d="M40,48c0,2.209-1.791,4-4,4h-8c-2.209,0-4-1.791-4-4V24c0-2.209,1.791-4,4-4h8c2.209,0,4,1.791,4,4V48z" fill="#FFFFFF"/></g><g><path d="M16.161,36c-2.722,0-3.483-1.865-1.69-4.145L28.741,13.71c1.793-2.28,4.726-2.28,6.519,0l14.269,18.146    c1.793,2.28,1.032,4.145-1.69,4.145H16.161z" fill="#FFFFFF"/></g></g><g id="Layer_2"/></svg>';
+        return '<span title="Slim Http Request">'.$this->icon.'</span>';
     }
 
     public function getPanel()
     {
-        return '<h1>Slim '.$this->ver['slim'].' Request:</h1><div>' . Dumper::toHtml($this->app->getContainer()->get('request')) . '</div>';
+        return '<h1>'.$this->icon.' Slim '.$this->ver['slim'].' Request:</h1><div style="overflow: auto; max-height: 600px;">' . Dumper::toHtml($this->app->getContainer()->get('request')) . '</div>';
     }
 }